Homework

+++ QA java course/src/main/java/com/javalectures/lecture10/Player.java
@@ -57,7 +57,7 @@ public class Player {
     }
 
     public float getOverallSkill(){
-        return (endurance + sprint + dribble + passing + shooting)/5;
+        return (float) (endurance + sprint + dribble + passing + shooting)/5;
     }
